Make the analyzer take a repository configuration object instead of a file
Having in mind that the analyzer API might be used in a server backend
without access to a user's file system, it is not possible to read a
repository configuraion file from an arbitrary file system location at
that point. So instead of the file, takle the repository configuration
object.

Signed-off-by: Sebastian Schuberth <sebastian.schuberth@bosch.io>